+++ webpack.config.base.js
@@ -109,12 +109,13 @@ const baseConfig = {
       },
       {
         test: /\.svg$/,
-        use: [{
-          loader: 'url-loader',
-          options: {
-            esModule: false,
-          },
-        }]
+        type: 'asset/inline',
+        resourceQuery: { not: /icon/ } // exclude built-in icons from stripes-components which are loaded as react components.
+      },
+      {
+        test: /\.svg$/,
+        resourceQuery: /icon/, // stcom icons use this query on the resource.
+        use: ['@svgr/webpack']
       },
     ],
   },